+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 0.6-pre3
 
+- **autoupdate**: Archive old versions of manifest when executing checkver/autoupdate
 - **Git**: Always use `--no-pager` option
 - **checkup**: Test Windows Defender exlusions only when executed with administrator privileges
 - Remove automatic config migration